# Improve performance of biter_raffle