--- a/maps/biter_battles_v2/biter_raffle.lua
+++ b/maps/biter_battles_v2/biter_raffle.lua
@@ -6,34 +6,33 @@ local math_max = math.max
 local function get_raffle_table(level)
     if level < 500 then
         return {
-            ['small-biter'] = 1000 - level * 1.75,
-            ['medium-biter'] = math_max(-250 + level * 1.5, 0), -- only this one can be negative for level < 500
-            ['big-biter'] = 0,
-            ['behemoth-biter'] = 0,
-        }
+            ['small-'] = 1000 - level * 1.75,
+            ['medium-'] = math_max(-250 + level * 1.5, 0), -- only this one can be negative for level < 500
+            ['big-'] = 0,
+            ['behemoth-'] = 0,
+        },
+            1000 - 1.75 * level + math_max(-250 + level * 1.5, 0)
     end
     if level < 900 then
         return {
-            ['small-biter'] = math_max(1000 - level * 1.75, 0), -- only this one can be negative for level < 900
-            ['medium-biter'] = 1000 - level,
-            ['big-biter'] = (level - 500) * 2,
-            ['behemoth-biter'] = 0,
-        }
+            ['small-'] = math_max(1000 - level * 1.75, 0), -- only this one can be negative for level < 900
+            ['medium-'] = 1000 - level,
+            ['big-'] = (level - 500) * 2,
+            ['behemoth-'] = 0,
+        },
+            math_max(1000 - level * 1.75, 0) + level
     end
     return {
-        ['small-biter'] = 0,
-        ['medium-biter'] = math_max(1000 - level, 0),
-        ['big-biter'] = (level - 500) * 2,
-        ['behemoth-biter'] = (level - 900) * 8,
-    }
+        ['small-'] = 0,
+        ['medium-'] = math_max(1000 - level, 0),
+        ['big-'] = (level - 500) * 2,
+        ['behemoth-'] = (level - 900) * 8,
+    },
+        math_max(1000 - level, 0) + 10 * level - 8200
 end
 
 local function roll(evolution_factor)
-    local raffle = get_raffle_table(math_floor(evolution_factor * 1000))
-    local max_chance = 0
-    for _, v in pairs(raffle) do
-        max_chance = max_chance + v
-    end
+    local raffle, max_chance = get_raffle_table(math_floor(evolution_factor * 1000))
     local r = math_random(0, math_floor(max_chance))
     local current_chance = 0
     for k, v in pairs(raffle) do