--- a/maps/biter_battles_v2/biter_raffle.lua
+++ b/maps/biter_battles_v2/biter_raffle.lua
@@ -10,8 +10,7 @@ local function get_raffle_table(level)
             ['medium-'] = math_max(-250 + level * 1.5, 0), -- only this one can be negative for level < 500
             ['big-'] = 0,
             ['behemoth-'] = 0,
-        },
-            1000 - 1.75 * level + math_max(-250 + level * 1.5, 0)
+        }
     end
     if level < 900 then
         return {
@@ -19,21 +18,19 @@ local function get_raffle_table(level)
             ['medium-'] = 1000 - level,
             ['big-'] = (level - 500) * 2,
             ['behemoth-'] = 0,
-        },
-            math_max(1000 - level * 1.75, 0) + level
+        }
     end
     return {
         ['small-'] = 0,
         ['medium-'] = math_max(1000 - level, 0),
         ['big-'] = (level - 500) * 2,
         ['behemoth-'] = (level - 900) * 8,
-    },
-        math_max(1000 - level, 0) + 10 * level - 8200
+    }
 end
 
 local function roll(evolution_factor)
-    local raffle, max_chance = get_raffle_table(math_floor(evolution_factor * 1000))
-    local r = math_random(0, math_floor(max_chance))
+    local raffle = get_raffle_table(math_floor(evolution_factor * 1000))
+    local r = math_random(0, math_floor(raffle[1] + raffle[2] + raffle[3] + raffle[4]))
     local current_chance = 0
     for k, v in pairs(raffle) do
         current_chance = current_chance + v