### Brief description of the changes:
As described in #477 

- Cut down on string concatenation
- Replaced `for` loop checking for every entry being positive with `math.max` checking only the entries that could get negative.
- Replaced `for` loop summing `rafle` with direct calls to it, as size of `raffle` table is fixed 

### Performance:

Before:
biter:  n=100000        total=170ms     avg=0.001700ms
spitter:        n=100000        total=169ms     avg=0.001690ms
worm:   n=100000        total=129ms     avg=0.001290ms
mixed:  n=100000        total=174ms     avg=0.001740ms 

After:
biter:  n=100000        total=104ms     avg=0.001040ms
spitter:        n=100000        total=104ms     avg=0.001040ms
worm:   n=100000        total=130ms     avg=0.001300ms
mixed:  n=100000        total=112ms     avg=0.001120ms

Around 35% - 40% improvement.

### Tested Changes:
- [x] I've tested the changes locally.
- [ ] I've not tested the changes.