@TheBigZet 2025-01-31
Possible room for improvement:
Changing condition  on line 88 from `<=` to `<` would allow to skip creation of keys with `0` value.