@cogito123 2025-01-31
I actually was working in the same area. I didn't manage to finish, but maybe you could incorporate some of my idea.
https://github.com/cogito123/Factorio-Biter-Battles/commit/19d5c920142bf77f8259b2f08dd1a3d964e84c50