@TheBigZet 2025-02-06
> I actually was working in the same area. I didn't manage to finish, but maybe you could incorporate some of my idea. [cogito123@19d5c92](https://github.com/cogito123/Factorio-Biter-Battles/commit/19d5c920142bf77f8259b2f08dd1a3d964e84c50)

Unfortunately, this way of creating arrays 
```
array = {
 [1] = ... , 
 [2] = ... 
}
```
does not guarantee creating an actual `array part`.
https://www.lua.org/gems/sample.pdf page 20.
I'm not sure this change actually improves performance, and I don't know if it can be fixed without losing readability